Resolve cyclic module imports as detected by LGTM.com
