+++ kopf/__init__.py
@@ -10,13 +10,15 @@ from kopf import (
     on,  # as a separate name on the public namespace
 )
 from kopf.config import (
-    configure,
     LOGLEVEL_INFO,  # deprecated
     LOGLEVEL_WARNING,  # deprecated
     LOGLEVEL_ERROR,  # deprecated
     LOGLEVEL_CRITICAL,  # deprecated
     EventsConfig,  # deprecated
     WorkersConfig,  # deprecated
+)
+from kopf.engines.logging import (
+    configure,
 )
 from kopf.engines.posting import (
     event,
