+++ tests/conftest.py
@@ -16,8 +16,7 @@ import pytest_mock
 
 import kopf
 from kopf.clients.auth import APIContext
-from kopf.config import configure
-from kopf.engines.logging import ObjectPrefixingFormatter
+from kopf.engines.logging import configure, ObjectPrefixingFormatter
 from kopf.engines.posting import settings_var
 from kopf.structs.configuration import OperatorSettings
 from kopf.structs.credentials import Vault, VaultKey, ConnectionInfo
